feat/MSSDK-1934: move types to components folders
