+++ src/appRedux/types/publisherConsentsSlice.types.ts
@@ -1,4 +1,4 @@
-import { Consent } from 'types/components/Consents.types';
+import { Consent } from 'types/Consents.types';
 
 export type PublisherConsentsInitialState = {
   publisherConsents: Consent[];
