+++ src/components/Capture/Capture.types.ts
