+++ src/components/Consents/Consents.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react';
 import Consent from 'components/Consents';
 import 'i18NextInit';
-import { Consent as ConsentType } from 'types/components/Consents.types';
+import { Consent as ConsentType } from 'types/Consents.types';
 import renderWithProviders from 'util/testHelpers';
 import { RootState } from 'appRedux/rootReducer';
 import { initialState } from 'appRedux/publisherConfigSlice';
