+++ src/components/Consents/Consents.types.ts
@@ -1,4 +1,4 @@
-import { Consent } from 'types/components/Consents.types';
+import { Consent } from 'types/Consents.types';
 
 export type ConsentsProps = {
   error?: string;
